Add docstring for the user_loader callback. (#360)
* Add some explanation for the user_loader callback.

Though this is not that significant, but I think this is necessary for 2 reasons: 1. it's hard to find clear answer for both why and how to correct this, especially for newbie; 2. this is my first pull request for a popular and great project, it matters! ;)

* Update login_manager.py

* Update login_manager.py

* Add a description into docstring.

* Update login_manager.py

Reason to delete the hint "add LoginManager.user_loader decorator...": basically, there're two groups of people who got this exception, who know how/where/why to add this callback, and others who don't. For first group, when they're notificated "user_loader" wasn't installed, they will immediately realize what's happening; for those who don't, they're more likely knowing few about this (like me), in this situation, "add ...decorator" won't help them figure this out, they need to deep into more details.
So, the exception message for this may should be "all or nothing".

* Update login_manager.py